feat: 목표 금액, 기간 페이지 분리
